+++ src/api/_credentialsFactory.ts
@@ -1,7 +1,6 @@
 import { AxiosResponse, default as axios } from 'axios';
 import { access, constants, readFileSync, writeFileSync } from 'fs';
 import { dirname } from 'path';
-import { sync } from '../../utils/mkdirp';
 import {
   AcountNumberString,
   Credentials,
@@ -10,7 +9,8 @@ import {
   IAccounts,
   ITime,
   QuestradeAPIOptions,
-} from '../libraries';
+} from '../core/libraries';
+import { sync } from '../utils/mkdirp';
 import { _axiosApiGet } from './_axiosApi';
 
 const _getServerTime = (credentials: Credentials) => async () =>
