+++ src/debugtesting.ts
@@ -1,7 +1,7 @@
 // tslint:disable: ordered-imports
-import { _credentialsFactory } from './core/api/_credentialsFactory';
+import { _credentialsFactory } from './api/_credentialsFactory';
 import { id0 } from './core/libraries';
-import { apiFunctions } from './core/private/_getDataFromApi';
+import { apiFunctions } from './private/_getDataFromApi';
 import { day } from './utils';
 export const log = console.log;
 const myRefreshToken = 'qd0AJcnxOGOKpXzvqzIbzKwgHy3Rm3XJ0';
